Clean Up Codebase and Remove Obsolete "wallets" Feature Flags and Implementations
