+++ changelog/remove-tokenized-cart-ece-flags
@@ -1,0 +1,4 @@
+Significance: minor
+Type: update
+
+Clean Up Codebase and Remove Obsolete "wallets" Feature Flags and Implementations
