+++ client/checkout/api/index.js
@@ -4,11 +4,7 @@
  * Internal dependencies
  */
 import { getConfig, getUPEConfig } from 'utils/checkout';
-import {
-	getExpressCheckoutConfig,
-	buildAjaxURL,
-	getExpressCheckoutAjaxURL,
-} from 'utils/express-checkout';
+import { getExpressCheckoutConfig, buildAjaxURL } from 'utils/express-checkout';
 import { getAppearance } from 'checkout/upe-styles';
 import { getAppearanceType } from '../utils';
 
@@ -324,115 +320,6 @@ export default class WCPayAPI {
 					throw new Error( error.statusText );
 				}
 			} );
-	}
-
-	/**
-	 * Updates cart with selected shipping option.
-	 *
-	 * @param {Object} shippingOption Shipping option.
-	 * @return {Promise} Promise for the request to the server.
-	 */
-	expressCheckoutECEUpdateShippingDetails( shippingOption ) {
-		return this.request(
-			getExpressCheckoutAjaxURL( 'ece_update_shipping_method' ),
-			{
-				security: getExpressCheckoutConfig( 'nonce' )?.update_shipping,
-				shipping_method: [ shippingOption.id ],
-				is_product_page:
-					getExpressCheckoutConfig( 'button_context' ) === 'product',
-			}
-		);
-	}
-
-	/**
-	 * Get cart items and total amount.
-	 *
-	 * @return {Promise} Promise for the request to the server.
-	 */
-	expressCheckoutECEGetCartDetails() {
-		return this.request(
-			getExpressCheckoutAjaxURL( 'ece_get_cart_details' ),
-			{
-				security: getExpressCheckoutConfig( 'nonce' )?.get_cart_details,
-			}
-		);
-	}
-
-	/**
-	 * Add product to cart from variable product page.
-	 *
-	 * @param {Object} productData Product data.
-	 * @return {Promise} Promise for the request to the server.
-	 */
-	expressCheckoutECEAddToCart( productData ) {
-		return this.request( getExpressCheckoutAjaxURL( 'add_to_cart' ), {
-			security: getExpressCheckoutConfig( 'nonce' )?.add_to_cart,
-			...productData,
-		} );
-	}
-
-	/**
-	 * Get selected product data from variable product page.
-	 *
-	 * @param {Object} productData Product data.
-	 * @return {Promise} Promise for the request to the server.
-	 */
-	expressCheckoutECEGetSelectedProductData( productData ) {
-		return this.request(
-			getExpressCheckoutAjaxURL( 'ece_get_selected_product_data' ),
-			{
-				security: getExpressCheckoutConfig( 'nonce' )
-					?.get_selected_product_data,
-				...productData,
-			}
-		);
-	}
-
-	/**
-	 * Submits shipping address to get available shipping options
-	 * from Express Checkout ECE payment method.
-	 *
-	 * @param {Object} shippingAddress Shipping details.
-	 * @return {Promise} Promise for the request to the server.
-	 */
-	expressCheckoutECECalculateShippingOptions( shippingAddress ) {
-		return this.request(
-			getExpressCheckoutAjaxURL( 'ece_get_shipping_options' ),
-			{
-				security: getExpressCheckoutConfig( 'nonce' )?.shipping,
-				is_product_page:
-					getExpressCheckoutConfig( 'button_context' ) === 'product',
-				...shippingAddress,
-			}
-		);
-	}
-
-	/**
-	 * Creates order based on Express Checkout ECE payment method.
-	 *
-	 * @param {Object} paymentData Order data.
-	 * @return {Promise} Promise for the request to the server.
-	 */
-	expressCheckoutECECreateOrder( paymentData ) {
-		return this.request( getExpressCheckoutAjaxURL( 'ece_create_order' ), {
-			_wpnonce: getExpressCheckoutConfig( 'nonce' )?.checkout,
-			...paymentData,
-		} );
-	}
-
-	/**
-	 * Pays for an order based on the Express Checkout payment method.
-	 *
-	 * @param {integer} order The order ID.
-	 * @param {Object} paymentData Order data.
-	 * @return {Promise} Promise for the request to the server.
-	 */
-	expressCheckoutECEPayForOrder( order, paymentData ) {
-		return this.request( getExpressCheckoutAjaxURL( 'ece_pay_for_order' ), {
-			_wpnonce: getExpressCheckoutConfig( 'nonce' )?.pay_for_order,
-			order,
-			...paymentData,
-		} );
 	}
 
 	initWooPay( userEmail, woopayUserSession ) {
