+++ client/checkout/api/test/index.test.js
@@ -3,11 +3,7 @@
  */
 import WCPayAPI from '..';
 import request from 'wcpay/checkout/utils/request';
-import {
-	buildAjaxURL,
-	getExpressCheckoutAjaxURL,
-	getExpressCheckoutConfig,
-} from 'wcpay/utils/express-checkout';
+import { buildAjaxURL } from 'wcpay/utils/express-checkout';
 import { getConfig } from 'wcpay/utils/checkout';
 
 jest.mock( 'wcpay/checkout/utils/request', () =>
@@ -143,20 +139,6 @@ describe( 'WCPayAPI', () => {
 			billing_email: 'test@example.com',
 		} );
 		expect( api.isWooPayRequesting ).toBe( false );
-	} );
-
-	test( 'express checkout pay for order is initialized correctly', async () => {
-		getExpressCheckoutAjaxURL.mockReturnValue( 'https://example.org/' );
-		getExpressCheckoutConfig.mockReturnValue( { pay_for_order: '1234' } );
-
-		const api = new WCPayAPI( {}, request );
-		await api.expressCheckoutECEPayForOrder( '12', { foo: 'bar' } );
-
-		expect( request ).toHaveBeenLastCalledWith( 'https://example.org/', {
-			_wpnonce: '1234',
-			order: '12',
-			foo: 'bar',
-		} );
 	} );
 
 	test( 'WooPay should not support global theme styles', async () => {
