+++ client/checkout/blocks/index.js
@@ -19,10 +19,6 @@ import PaymentMethodLabel from './payment-method-label';
 import request from '../utils/request';
 import enqueueFraudScripts from 'fraud-scripts';
 import {
-	expressCheckoutElementApplePay,
-	expressCheckoutElementGooglePay,
-} from '../../express-checkout/blocks';
-import {
 	tokenizedExpressCheckoutElementApplePay,
 	tokenizedExpressCheckoutElementGooglePay,
 } from 'wcpay/tokenized-express-checkout/blocks';
@@ -161,17 +157,12 @@ if ( getUPEConfig( 'isWooPayEnabled' ) ) {
 }
 
 if ( getUPEConfig( 'isPaymentRequestEnabled' ) ) {
-	if ( getUPEConfig( 'isTokenizedCartEceEnabled' ) ) {
-		registerExpressPaymentMethod(
-			tokenizedExpressCheckoutElementApplePay( api )
-		);
-		registerExpressPaymentMethod(
-			tokenizedExpressCheckoutElementGooglePay( api )
-		);
-	} else {
-		registerExpressPaymentMethod( expressCheckoutElementApplePay( api ) );
-		registerExpressPaymentMethod( expressCheckoutElementGooglePay( api ) );
-	}
+	registerExpressPaymentMethod(
+		tokenizedExpressCheckoutElementApplePay( api )
+	);
+	registerExpressPaymentMethod(
+		tokenizedExpressCheckoutElementGooglePay( api )
+	);
 }
 window.addEventListener( 'load', () => {
 	enqueueFraudScripts( getUPEConfig( 'fraudServices' ) );
