--- client/express-checkout/blocks/components/express-checkout-container.js
@@ -1,42 +1,0 @@
-/**
- * External dependencies
- */
-import { useMemo } from 'react';
-import { Elements } from '@stripe/react-stripe-js';
-
-/**
- * Internal dependencies
- */
-import ExpressCheckoutComponent from './express-checkout-component';
-import {
-	getExpressCheckoutButtonAppearance,
-	getExpressCheckoutData,
-} from 'wcpay/express-checkout/utils';
-import '../express-checkout-element.scss';
-
-const ExpressCheckoutContainer = ( props ) => {
-	const { api, billing, buttonAttributes, isPreview } = props;
-
-	const stripePromise = useMemo( () => {
-		return api.loadStripeForExpressCheckout();
-	}, [ api ] );
-
-	const options = {
-		mode: 'payment',
-		paymentMethodCreation: 'manual',
-		amount: ! isPreview ? billing.cartTotal.value : 10,
-		currency: ! isPreview ? billing.currency.code.toLowerCase() : 'usd',
-		appearance: getExpressCheckoutButtonAppearance( buttonAttributes ),
-		locale: getExpressCheckoutData( 'stripe' )?.locale ?? 'en',
-	};
-
-	return (
-		<div style={ { minHeight: '40px' } }>
-			<Elements stripe={ stripePromise } options={ options }>
-				<ExpressCheckoutComponent { ...props } />
-			</Elements>
-		</div>
-	);
-};
-
-export default ExpressCheckoutContainer;
