--- client/express-checkout/blocks/hooks/use-express-checkout.js
@@ -1,115 +1,0 @@
-/**
- * External dependencies
- */
-import { useCallback } from '@wordpress/element';
-import { useStripe, useElements } from '@stripe/react-stripe-js';
-
-/**
- * Internal dependencies
- */
-import {
-	displayLoginConfirmation,
-	getExpressCheckoutButtonStyleSettings,
-	getExpressCheckoutData,
-	normalizeLineItems,
-} from 'wcpay/express-checkout/utils';
-import {
-	onAbortPaymentHandler,
-	onCancelHandler,
-	onClickHandler,
-	onCompletePaymentHandler,
-	onConfirmHandler,
-	onReadyHandler,
-} from 'wcpay/express-checkout/event-handlers';
-
-export const useExpressCheckout = ( {
-	api,
-	billing,
-	shippingData,
-	onClick,
-	onClose,
-	setExpressPaymentError,
-} ) => {
-	const stripe = useStripe();
-	const elements = useElements();
-
-	const buttonOptions = getExpressCheckoutButtonStyleSettings();
-
-	const onCancel = () => {
-		onCancelHandler();
-		onClose();
-	};
-
-	const completePayment = ( redirectUrl ) => {
-		onCompletePaymentHandler();
-		window.location = redirectUrl;
-	};
-
-	const abortPayment = ( onConfirmEvent, message ) => {
-		onConfirmEvent.paymentFailed( { reason: 'fail' } );
-		setExpressPaymentError( message );
-		onAbortPaymentHandler();
-	};
-
-	const onButtonClick = useCallback(
-		( event ) => {
-			// If login is required for checkout, display redirect confirmation dialog.
-			if ( getExpressCheckoutData( 'login_confirmation' ) ) {
-				displayLoginConfirmation( event.expressPaymentType );
-				return;
-			}
-
-			const options = {
-				lineItems: normalizeLineItems( billing?.cartTotalItems ),
-				emailRequired: true,
-				shippingAddressRequired: shippingData?.needsShipping,
-				phoneNumberRequired:
-					getExpressCheckoutData( 'checkout' )?.needs_payer_phone ??
-					false,
-				shippingRates: shippingData?.shippingRates[ 0 ]?.shipping_rates?.map(
-					( r ) => {
-						return {
-							id: r.rate_id,
-							amount: parseInt( r.price, 10 ),
-							displayName: r.name,
-						};
-					}
-				),
-				allowedShippingCountries: getExpressCheckoutData( 'checkout' )
-					.allowed_shipping_countries,
-			};
-
-			// Click event from WC Blocks.
-			onClick();
-			// Global click event handler from WooPayments to ECE.
-			onClickHandler( event );
-			event.resolve( options );
-		},
-		[
-			onClick,
-			billing.cartTotalItems,
-			shippingData.needsShipping,
-			shippingData.shippingRates,
-		]
-	);
-
-	const onConfirm = async ( event ) => {
-		onConfirmHandler(
-			api,
-			stripe,
-			elements,
-			completePayment,
-			abortPayment,
-			event
-		);
-	};
-
-	return {
-		buttonOptions,
-		onButtonClick,
-		onConfirm,
-		onReady: onReadyHandler,
-		onCancel,
-		elements,
-	};
-};
