--- client/express-checkout/utils/test/index.ts
@@ -1,44 +1,0 @@
-/**
- * Internal dependencies
- */
-import {
-	WCPayExpressCheckoutParams,
-	getErrorMessageFromNotice,
-	getExpressCheckoutData,
-} from '..';
-
-describe( 'Express checkout utils', () => {
-	test( 'getExpressCheckoutData returns null for missing option', () => {
-		expect(
-			getExpressCheckoutData(
-				// Force wrong usage, just in case this is called from JS with incorrect params.
-				'does-not-exist' as keyof WCPayExpressCheckoutParams
-			)
-		).toBeNull();
-	} );
-
-	test( 'getExpressCheckoutData returns correct value for present option', () => {
-		// We don't care that the implementation is partial for the purposes of the test, so
-		// the type assertion is fine.
-		window.wcpayExpressCheckoutParams = {
-			ajax_url: 'test',
-		} as WCPayExpressCheckoutParams;
-
-		expect( getExpressCheckoutData( 'ajax_url' ) ).toBe( 'test' );
-	} );
-
-	test( 'getErrorMessageFromNotice strips formatting', () => {
-		const notice = '<p><b>Error:</b> Payment failed.</p>';
-		expect( getErrorMessageFromNotice( notice ) ).toBe(
-			'Error: Payment failed.'
-		);
-	} );
-
-	test( 'getErrorMessageFromNotice strips scripts', () => {
-		const notice =
-			'<p><b>Error:</b> Payment failed.<script>alert("hello")</script></p>';
-		expect( getErrorMessageFromNotice( notice ) ).toBe(
-			'Error: Payment failed.alert("hello")'
-		);
-	} );
-} );
