+++ webpack/shared.js
@@ -19,7 +19,6 @@ module.exports = {
 				'./client/checkout/woopay/direct-checkout/index.js',
 			cart: './client/cart/index.js',
 			checkout: './client/checkout/classic/event-handlers.js',
-			'express-checkout': './client/express-checkout/index.js',
 			'tokenized-express-checkout':
 				'./client/tokenized-express-checkout/index.js',
 			'subscription-edit-page': './client/subscription-edit-page.js',
