+++ woocommerce-payments.php
@@ -160,9 +160,7 @@ function wcpay_init() {
 	 * Check https://github.com/Automattic/woocommerce-payments/issues/4759
 	 */
 	\WCPay\WooPay\WooPay_Session::init();
-	if ( WC_Payments_Features::is_tokenized_cart_ece_enabled() ) {
-		( new WC_Payments_Payment_Request_Session() )->init();
-	}
+	( new WC_Payments_Payment_Request_Session() )->init();
 }
 
 // Make sure this is run *after* WooCommerce has a chance to initialize its packages (wc-admin, etc). That is run with priority 10.
